pyproject.toml
[tool.poetry]
name = "advancedragingest"
version = "0.1.0"
description = "PDF processing tool for content extraction and annotation"
authors = ["Prof. Tiran Dagan <your.email@fdu.edu>"]
license = "Proprietary"
readme = "README.md"
packages = [
    { include = "helpers" },
    { include = "supabase" }
]

[tool.poetry.dependencies]
python = ">=3.9,<4.0"
PyMuPDF = "^1.23.0"
PyPDF2 = "^3.0.0"
pdf2image = "^1.16.3"
pdfminer-six = "^20221105"
rich = "^13.7.0"
pyyaml = "^6.0.1"
unstructured = "^0.10.30"
unstructured-inference = "^0.7.21"
Pillow = "^10.0.0"
tqdm = "^4.66.1"
pytesseract = "^0.3.10"
numpy = "^1.24.0"
langchain = "0.1.0"
supabase = "2.0.3"
langchain-community = "0.0.16"
openai = "1.12.0"
langchain-openai = "0.0.5"
python-dotenv = "^1.0.1"
inquirer = "^3.4.0"
pandas = "^2.2.3"
tabulate = "^0.9.0"

[tool.poetry.group.dev.dependencies]
pytest = "^7.4.0"
black = "^23.7.0"
isort = "^5.12.0"
flake8 = "^6.1.0"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"

[tool.black]
line-length = 88
target-version = ['py38']
include = '\.pyi?$'

[tool.isort]
profile = "black"
multi-line-output = 3

